🚀 Improve: Select URL randomly at the very end
The URL is now randomly selected at the very end, not while reading the
ingredient_category_url.csv file.

Having a random attribute in Ingredient makes it harder to write tests,
especially because reading (the CSV) with multiple URLs and assembling the Ingredient
instance is an important feature.
Now, the behavior is far more deterministic since and randomness is
moved to the very end.

Nevertheless, an attribute for the chosen URL is necessary to print the
final shopping list including the URL.